requirements.txt

# Python version : 3.12.9

streamlit
pillow
